--- src/sections/services/api/Services.js
@@ -1,10 +1,0 @@
-const yaml = require('js-yaml');
-const fs   = require('fs');
-
-// Get document, or throw exception on error
-try {
-  const doc = yaml.load(fs.readFileSync('../ServicesData.yml', 'utf8'));
-  console.log(doc);
-} catch (e) {
-  console.log(e);
-}
