+++ src/sections/services/servicesData.js
@@ -1,0 +1,39 @@
+import ProductScoping from "../../assets/illustration/principles/product_scoping.svg";
+import Design from "../../assets/illustration/principles/design.svg";
+import Mvp from "../../assets/illustration/principles/mvp.svg";
+import DevelopmentPartnership from "../../assets/illustration/principles/development_partnership.svg";
+
+export const serviceData = [
+  {
+    id: 1,
+    title: "product scoping",
+    color: "bg-brand-background-300",
+    image: ProductScoping,
+    description:
+      "You wouldn't build a house without a blueprint, so why launch a product without a solid product scope? Unleash your curiosity and learn more about our product scoping processes.",
+  },
+  {
+    id: 2,
+    title: "designs",
+    color: "bg-brand-background-400",
+    image: Design,
+    description:
+      "Being a crucial part of product development, a well-designed software product is not only aesthetically pleasing but also intuitive and easy to use. A great design can make a product stand out in a crowded market. Explore our product design approach.",
+  },
+  {
+    id: 3,
+    title: "MVP Development",
+    color: "bg-brand-background-300",
+    image: Mvp,
+    description:
+      "A key benefit of building an MVP is that, it helps you to reduce the risk of product failure. By testing the MVP with real users, you can obtain feedback, identify potential issues, and make informed decisions about the product's development. Dive into our MVP development process.",
+  },
+  {
+    id: 4,
+    title: "Software Development partnership",
+    color: "bg-brand-background-400",
+    image: DevelopmentPartnership,
+    description:
+      "Partnering with us for development offers two significant benefits. Firstly, we provide the ability to experiment, prototype, and validate new ideas. Secondly, we offer our extensive skill set to build your initial MVP and scale it to new heights. Together, we can reach to PMF that you aim for.",
+  },
+];
